move icons list to constants
