+++ frontend/src/constants/icons.ts
@@ -1,0 +1,25 @@
+// Enum for badge icon options
+export enum Icon {
+  FIVE_STAR = "five_star",
+  FOUR_STAR = "four_star",
+  GROUP = "group",
+  HEART = "heart",
+  HOME = "home",
+  BABY = "baby",
+  WINGS = "wings",
+  FLOWER = "flower",
+  MONEY = "money",
+  GEMSTONE = "gemstone",
+  DIAMOND = "diamond",
+  PENCIL = "pencil",
+  TOOL = "tool",
+}
+
+export const iconList = [
+  Icon.FIVE_STAR,
+  Icon.GROUP,
+  Icon.HEART,
+  Icon.HOME,
+  Icon.BABY,
+  Icon.WINGS,
+];
